+++ NEWS.md
@@ -2,7 +2,8 @@
 
 ## Bug fixes
 
-* Fix aliasing detection in `sort!`
+* Fix aliasing detection in `sort!` (now only identical columns passing `===`
+  test are considered aliases)
   ([#2981](https://github.com/JuliaData/DataFrames.jl/issues/2981))
 
 # DataFrames.jl v1.3.1 Patch Release Notes